requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
drf-yasg>=1.20.0,<2.0
django-cors-headers>=3.7,<4.0
matplotlib>=3.4,<4.0
bcrypt